refactor: TodoView -> TodoListVC와 병합,  TodoViewModel 삭제
